refactor: move artifacts utilities from artifacts.py to storage module and remove redundant core/types.py module
