+++ packages/devqubit-engine/src/devqubit_engine/bundle/pack.py
@@ -42,7 +42,7 @@ from pathlib import Path
 from typing import Any
 
 from devqubit_engine.storage.factory import create_registry, create_store
-from devqubit_engine.storage.protocols import ObjectStoreProtocol, RegistryProtocol
+from devqubit_engine.storage.types import ObjectStoreProtocol, RegistryProtocol
 from devqubit_engine.utils.common import utc_now_iso
