+++ packages/devqubit-engine/src/devqubit_engine/circuit/extractors.py
@@ -15,18 +15,15 @@ import logging
 import re
 from typing import TYPE_CHECKING
 
-from devqubit_engine.artifacts import (
-    find_artifact,
-    load_artifact_bytes,
-)
 from devqubit_engine.circuit.models import SDK, CircuitData, CircuitFormat
 from devqubit_engine.core.record import RunRecord
-from devqubit_engine.storage.protocols import ObjectStoreProtocol
-from devqubit_engine.uec.types import ArtifactRef
+from devqubit_engine.storage.artifacts.browse import find_artifact
+from devqubit_engine.storage.artifacts.io import load_artifact_bytes
+from devqubit_engine.storage.types import ArtifactRef, ObjectStoreProtocol
 
 
 if TYPE_CHECKING:
-    from devqubit_engine.uec.envelope import ExecutionEnvelope
+    from devqubit_engine.uec.models.envelope import ExecutionEnvelope
 
 
 logger = logging.getLogger(__name__)
