+++ packages/devqubit-engine/src/devqubit_engine/cli/admin.py
@@ -220,8 +220,8 @@ def baseline_group() -> None:
 def baseline_set(ctx: click.Context, project: str, run_id: str) -> None:
     """Set baseline run for a project."""
     from devqubit_engine.core.config import Config
+    from devqubit_engine.storage.errors import RunNotFoundError
     from devqubit_engine.storage.factory import create_registry
-    from devqubit_engine.storage.protocols import RunNotFoundError
 
     root = root_from_ctx(ctx)
     config = Config(root_dir=root)
