+++ packages/devqubit-engine/src/devqubit_engine/compare/diff.py
@@ -19,10 +19,6 @@ from numbers import Real
 from pathlib import Path
 from typing import Any, Iterator
 
-from devqubit_engine.artifacts import (
-    get_artifact_digests,
-    get_counts,
-)
 from devqubit_engine.bundle.reader import Bundle, is_bundle_path
 from devqubit_engine.circuit.extractors import extract_circuit
 from devqubit_engine.circuit.summary import (
@@ -42,16 +38,19 @@ from devqubit_engine.compare.results import (
 )
 from devqubit_engine.core.config import Config, get_config
 from devqubit_engine.core.record import RunRecord
+from devqubit_engine.storage.artifacts.browse import get_artifact_digests
+from devqubit_engine.storage.artifacts.counts import get_counts
 from devqubit_engine.storage.factory import create_registry, create_store
-from devqubit_engine.storage.protocols import ObjectStoreProtocol, RegistryProtocol
-from devqubit_engine.uec.calibration import DeviceCalibration
-from devqubit_engine.uec.device import DeviceSnapshot
-from devqubit_engine.uec.envelope import ExecutionEnvelope
-from devqubit_engine.uec.resolver import (
-    resolve_envelope,
+from devqubit_engine.storage.types import (
+    ArtifactRef,
+    ObjectStoreProtocol,
+    RegistryProtocol,
 )
-from devqubit_engine.uec.result import canonicalize_bitstrings
-from devqubit_engine.uec.types import ArtifactRef
+from devqubit_engine.uec.models.calibration import DeviceCalibration
+from devqubit_engine.uec.models.device import DeviceSnapshot
+from devqubit_engine.uec.models.envelope import ExecutionEnvelope
+from devqubit_engine.uec.models.result import canonicalize_bitstrings
+from devqubit_engine.uec.resolution.resolve import resolve_envelope
 from devqubit_engine.utils.distributions import (
     compute_noise_context,
     normalize_counts,
