+++ packages/devqubit-engine/src/devqubit_engine/compare/verify.py
@@ -19,7 +19,7 @@ from devqubit_engine.compare.diff import diff_runs
 from devqubit_engine.compare.results import ProgramMatchMode, VerifyResult
 from devqubit_engine.compare.verdict import build_verdict
 from devqubit_engine.core.record import RunRecord
-from devqubit_engine.storage.protocols import ObjectStoreProtocol
+from devqubit_engine.storage.types import ObjectStoreProtocol
 
 
 logger = logging.getLogger(__name__)
