+++ packages/devqubit-engine/src/devqubit_engine/storage/gc.py
@@ -18,7 +18,7 @@ from dataclasses import dataclass, field
 from datetime import datetime, timedelta, timezone
 from typing import Any
 
-from devqubit_engine.storage.protocols import (
+from devqubit_engine.storage.types import (
     ObjectStoreProtocol,
     RegistryProtocol,
     RunSummary,
