+++ packages/devqubit-engine/src/devqubit_engine/storage/local.py
@@ -25,13 +25,13 @@ from pathlib import Path
 from typing import Any, Iterator
 
 from devqubit_engine.core.record import RunRecord
-from devqubit_engine.storage.protocols import (
+from devqubit_engine.storage.types import (
+    ArtifactRef,
     BaselineInfo,
     ObjectNotFoundError,
     RunNotFoundError,
     RunSummary,
 )
-from devqubit_engine.uec.types import ArtifactRef
 from devqubit_engine.utils.common import utc_now_iso
