+++ packages/devqubit-engine/src/devqubit_engine/storage/remote_gcs.py
@@ -21,13 +21,13 @@ import re
 from typing import Any, Dict, Iterator, List
 
 from devqubit_engine.core.record import RunRecord
-from devqubit_engine.storage.protocols import (
+from devqubit_engine.storage.types import (
+    ArtifactRef,
     BaselineInfo,
     ObjectNotFoundError,
     RunNotFoundError,
     RunSummary,
 )
-from devqubit_engine.uec.types import ArtifactRef
 from devqubit_engine.utils.common import utc_now_iso
