+++ packages/devqubit-engine/tests/test_circuit.py
@@ -29,7 +29,7 @@ from devqubit_engine.circuit.summary import (
     CircuitSummary,
     summarize_circuit_data,
 )
-from devqubit_engine.uec.types import ArtifactRef
+from devqubit_engine.storage.types import ArtifactRef
 
 
 def sdk_available(sdk_name: str) -> bool:
