+++ packages/devqubit-qiskit/src/devqubit_qiskit/envelope.py
@@ -14,21 +14,22 @@ import logging
 from typing import Any
 
 from devqubit_engine.core.run import Run
-from devqubit_engine.uec.device import DeviceSnapshot
-from devqubit_engine.uec.envelope import ExecutionEnvelope
-from devqubit_engine.uec.execution import ExecutionSnapshot
-from devqubit_engine.uec.program import (
+from devqubit_engine.storage.types import ArtifactRef
+from devqubit_engine.uec.models.device import DeviceSnapshot
+from devqubit_engine.uec.models.envelope import ExecutionEnvelope
+from devqubit_engine.uec.models.execution import ExecutionSnapshot
+from devqubit_engine.uec.models.program import (
     ProgramArtifact,
     ProgramSnapshot,
     TranspilationInfo,
+    TranspilationMode,
 )
-from devqubit_engine.uec.result import (
+from devqubit_engine.uec.models.result import (
     CountsFormat,
     ResultError,
     ResultItem,
     ResultSnapshot,
 )
-from devqubit_engine.uec.types import ArtifactRef, TranspilationMode
 from devqubit_engine.utils.common import utc_now_iso
 from devqubit_engine.utils.serialization import to_jsonable
 from devqubit_qiskit.results import extract_result_metadata, normalize_result_counts
